Organize function / code into separate file Incorporate from & import statements to link files
